third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-time/timeline-time-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Test nesting of time/timeEnd records on Timeline


Running: testSimpleConsoleTime

Running: testNestedConsoleTime

Running: testUnbalancedConsoleTime

Running: testConsoleTimeWithoutConsoleTimeEnd
--------> Timestamp: Foo
--------> Timestamp: Bar
--------> Timestamp: Baz
